Add very basic power and gamma graph functions
The gamma graph allows users to have variables for edge weights. It uses
a function factory, which may not be the best way to do it, but it works
for now

The power function does not use anything as complex as the random number
generator from `gMCP::calcPower()`. It just generates p-values with
`mvtnorm::pmvnorm()` and tests the graph with each set generated. I
don't know if it's sufficiently random