Add C++ emulation of gMCP::testGraph
I tried to copy over gMCP's C code for sequential Bonferroni testing,
and it's working. But the all-C++ testing is even slower than the all-R
testing, and when I put it in a power function, it crashes. The best
option so far has been to re-write `zero_node_fast()` in C++ and just
use that small piece inside the rest using R. That reduces runtime by
about a half, but still doesn't come close to gMCP speed.

I'd like to look into using the C code from gMCP directly. Integrating
that and making the code around it a little more efficient could improve
performance more yet.