Improve Simes testing performance by pre-sorting hypotheses
For a given p-vector, a graph's hypotheses can be sorted before doing
any testing. The sort order holds across all intersections & groups
because p-values are not affected by node deletion. This yields ~9%
improvement on power simulation speed

It's possible there is more efficiency to gain in power simulations by
storing every possible sort order and indexing in to that list. But this
would be quite complex, and it might not actually gain speed, especially
for larger graphs