Improve parametric power performance by pre-calculating critical values
Because the critical value method doesn't depend on p-values, it's
possible to run generate_weights() and calculate all the critical values
at the start of the simulation. Then testing complexity amounts to
little more than Bonferroni. This yields a >30% time improvement from
only pre-calculating weights for a graph of size 5. Moving the
multivariate normal work from the testing function to the power function
also makes the testing portion a good candidate for re-writing in C++
for further speed improvement