Add fast `_v` versions of para/bonf testing
The essence of these versions comes from vectorizinge the testing, and
the processing of the test results. Rather than calculating adjusted
p-values for each group/intersection, p-values are compared directly to
the alpha times the weight matrix (times the critical values if
applicable, also stored as a matrix). Processing the test results is
then simply a column-wise max of the test results per intersection,
multiplied by the intersection matrix.

Simes testing will be a little more complicated, but it may be able to
follow a similar pattern. Coming out of something like `add_critical2()`, weights could
be ordered in the correct way to use `cumsum()` on them. But this would
still need to be run for every simulation, rather than all at the
beginning. And then the actual testing would be faster at least. This
might shave off the couple seconds that we're looking for