Update stat med 2010 table 1 imitation
In the process of updating I discovered a poor design aspect, where
passing a vector instead of a list as a group specification caused the
test specification to be ignored. I added a validation for it and
corresponding unit test.

Next step is to clean up the testing vignette and possibly make a
separte one for power. Then a power report class, most likely