Correct parametric power bugs and C++ sequential function
It took some fine-tuning on the upper limit of the critical value
integral to find something satisfactory. Too high and it can cause
errors when there are very small weights, too low and it can fail to
have a zero in the range. It's now been set to size of the graph plus
one, which should be the largest it ever needs to be

I also cleaned up some R CMD check notes and warnings, restored the
graph coercion functions I accidentally deleted, and