Complete first draft of case study vignette
With this update, the case study vignette is done-ish. It needs some
proofreading and re-wording I'm sure. There are also some points that
need more input from Dong, where I just don't have the domain knowledge
for them yet.

We also need to decide if we want to include plots at this point or not